requirements.txt
pandas>=2.0.0
reportlab>=4.0.0
matplotlib>=3.7.0
spotipy>=2.23.0
tqdm>=4.65.0
python-dateutil>=2.8.2